requirements.txt
python-dotenv
schedule
requests
pytz
psutil
